i realized that my section observer state was causing issues on larger screens so i fixed it.
Long story short: returning sections individually was causing issues with
knowing whether or not a section was inview at the same time as another section.
Ofcourse this caused issues with larger screens and making sure things show
when they should. so I made a small update to state that registers the
section as a key with a value of bool. now i can track all sections at
once and not have to worry if multiple are inview.